coll/base,tuned: introduce allgather_reduce allreduce algorithm
This patch introduces a new allreduce algorithm implemented as an
allgather followed by local reduction.

The change is motivated by the longer latency of tcp/EFA traffic.
Current allreduce algorithms require a round trip to and from a selected
root process. This algorithm avoids the round trip over network and
therefore reduces total latency.

However, this communication pattern is not scalable for large
communicators, and should only be used for inter-node allreduce.

Co-authored-by: Matt Koop <mkoop@amazon.com>
Co-authored-by: Wenduo Wang <wenduwan@mazon.com>

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>