Revert "pml/cm: fix buffer usage in MCA_PML_CM_HVY_SEND_REQUEST_BSEND_ALLOC()"
This reverts commit d71fe934c435a77480fabd4b21aeffce33bc770f.

The revert fixes a bug revealed by mtt ibm test suite. The send buffer,
instead of the attached user buffer, was used for MPI_Bsend.

This violates the MPI_Bsend semantic and makes it unsafe to reuse the
send buffer after the function returns.

The revert is also needed for 4.x

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>