fbtl/posix: fix data-sieving calculations
as part of introducing atomicity support for ompi v5.0, we also tried to improve the robustness in some file I/O routines.
Unfortunately, this also introduced a bug since ret_code returned by a function does not necessarily contain the number of
bytes read or written,
but could contain the last value (e.g. 0). The value was however used in a subsequent calculation and we ended not copying
data out of the temporary buffer used in the data sieving at all.

This commit also simplifies some of the logic in the while loop, no need to retry to read past the end of the file
multiple times.

Fixes issue #11917

Code was tested with the reproducer provided as part of the issue, our internal testsuite, and the hdf5-1.4.2 testsuite, all tests pass.

Signed-off-by: Edgar Gabriel <edgar.gabriel@amd.com>
(cherry picked from commit fb3b68f4b8dbc34958650db072fc2462c77024a8)