dlopen: fix configury issue
Turns out most of the time we are getting lucky and some opal dependency needs dlopen lib.
But there are certain types of builds, in particular a common one when building Open MPI
within spack that that is not the case.

In this event the opal_pal lib fails to build owing to missing dlopen/dlsym, etc.

This issue currently prevents building typical spack variants of open mpi 5.0.0

Related to https://github.com/spack/spack/pull/40725

Signed-off-by: Howard Pritchard <howardp@lanl.gov>