docs: replace many specific references to MPI-1
Where relevant, replace specific references to MPI-1 to generic
references to the MPI standard (linked out to the MPI Forum web site)
with either a chapter/section number that is still correct, or, if the
section/chapter number has changed over time, use the section/chapter
title (which is correct in both the old version of the MPI standard
and the current [MPI-4.1] version).

In a few places, this also carried a few trivial RST formatting
updates, and sections were re-flowed after adding the link to the MPI
Forum web site, but in general, no other text or content changed.

Signed-off-by: Jeff Squyres <jeff@squyres.com>