ORTE: a prefix of "/" is ok
This is a backport of https://github.com/openpmix/prrte/pull/1848 to
ORTE.  A prefix of "/" is not an error; there's no need to emit a
message and fail.  We want to ensure to not pass a double slash in the
ssh module to the remote host (i.e., the prefix of "/" followed by
"/<bindir>"), so make sure to handle that properly.

Signed-off-by: Jeff Squyres <jeff@squyres.com>