Github issue: add an "ask a question" template
We've had a recent spate of users wholly disregarding the template
when submitting a new github issue (i.e., they delete the template and
just ask a question, without all the useful information that the
template asks for).

For users who just want to ask a question, let's give them an explicit
template just for that (so that they don't just choose "bug report"
and then erase the template because they think the template won't
apply to them).

Signed-off-by: Jeff Squyres <jeff@squyres.com>