sessions: fix a problem with fortran interface
For fortran interfaces for MPI_Session_get_nth_pset,
the input character array in which to return the process set names
was not being properly handled to trim out potential junk
characers in the input array.

This is easiest to observe when the user is supplying an
input character array of MPI_MAX_PSET_NAME_LEN size.

Related to #12084.

Also do some fix-up of capitalization and indention.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>