fs/lustre: recognize soft-links on file open
the llapi_file_get_stripe() function does not accept as an input file a
soft-link. Check therefore whether the filename provided in File_open
is a soft-link, retrieve the real file name in case it is, make sure
it is also on a Lustre file system, and use the real filename in the
llapi_file_get_stripe() function call instead.

Fixes issue #12141

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>