fbtl/posix: prepare component for new request types
prepare the component to support more methods than aio_read/aio_write for
aysnchronous file I/O. For this, we need to distinguish between variables
in a structure that are required for aio operations, and once that are
generically required to manage the status of a parallel I/O request object.
Move the aio specific files into a structure that is part of a union.

Two more minor changes:
 - untabify some files.
 - replace a 'return OMPI_ERROR' by 'return false' in a function that is supposed
   to return a bool.

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>

asdf