move include of sys/mount back to where it was
Turns out the sys/mount.h can be tricky to use
because various other system include files sometimes redefine
some of the symbols in this include file, leading
to compile failures.

Careful ordering of include files may solve the problem,
but here it simpler just to move the include of sys/mount.h
back to its original location.

related to #12181

Signed-off-by: Howard Pritchard <howardp@lanl.gov>