Update Github issue templates
Update the sample version numbers that we list in the Github issue
templates to be v4.1.6 and v5.0.2 (vs. the v3.0.x and v4.0.x versions
we were citing before).  We don't need to update these every time we
release, but it is good to show at least the current generation of
major.minor series that are supported.

Signed-off-by: Jeff Squyres <jeff@squyres.com>