Fix incorrect TCP connections.
If nodes have the same IP addresses (for containers or other purposes) and
these addresses get published as part of the modex, a remote peer might try to
use one of the addresses to connect. As both nodes have the same IP, there are
several cases:
- the "remote" port is not used by an OMPI process locally, the connection is
  refused or it timeouts. This is the "nicest" outcome, as a new IP will be
  used resulting in a successful connection and the continuation of the
  application.
- the "remote" port is used by another OMPI process on the local node. A
  connection will be established but the incorrect guid will be exchanged
  leading to complaints, connection dropped and/or deadlocks.
- the "remote" port is used by this process, basically resulting in a
  connection-to-self. Bad things happen, as we don't support TCP connections to
  self. Some output messages are generated, but the outcome is most likely a
  deadlock.

Up to now, users were expected to exclude such interfaces from the accepted
interfaces, but this patch removes this need. If we discover a local IP as part
of the IP list of a remote peer, we drop it and never try to use it. This does
not apply to local processes, so we can still use these interfaces for node
level communications (which will work as we will connect to the correct port
according to the destination process).

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>