coll/han: Add alltall to HAN
Add two Alltoall algorithms to coll/han.  Both algorithms use the same
communication pattern.  Each rank on one host is assigned a single
partner on a remote host and vice versa.  Then the rank collects all
the data its partner will need to receive from it's host, and sends it
in one large send, and likewise receives it's data in one large recv,
then cycles to the next host.

The two algorithms are:
- mca_coll_han_alltoall_using_allgather: gathering data is done once
  and each rank has a copy of all local data.  Only recommended for
  small message sizes.
- mca_coll_han_alltoall_using_smsc: ranks use smsc module to
  direct-map local memory before copying into a packed send buffer.
  Currently only the XPMEM-based smsc module supports this operation.

Signed-off-by: Luke Robison <lrbison@amazon.com>