alltoallv: check send and recv datatype sizes separately
This is a followup to #12198. The previous change introduced a
bug that skips send and recv if either datatype is of 0-size.
This is wrong because MPI_Alltoallv allows each process to use
a unique datatype.

The correct solution is to check send and recv datatype size
separately and skip send and recv accordingly. This includes the
pairwise algorithm which should not send/recv 0-count data.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>