CI: Bot to mark abandoned Github Issues stale/closed
When an Open MPI community member assigns the label "State: Awaiting
user information" to a Github Issue, if someone without write
permissions to the repo doesn't reply in 14 days, mark the Github
Issue stale.  If no reply is received in a further 14 days, close the
Github Issue.

See lengthy comment in close-stale-issues.yaml for a more detailed
explanation.

Signed-off-by: Jeff Squyres <jeff@squyres.com>