Eliminate singleton session directories
If the host environment fails to provide us with a session
directory, then don't attempt to create one ourselves. The
only users of it are the shared memory subsystems (for
backing files) and the opal_output subsystem (when file
output is requested).

The sm btl knows how to deal with the scenario of no session
directories. The smcuda btl does _not_ know how to deal with
it, but someone can update it if/when they care.

The opal_output subsystem also knows how to deal with it. So
the going opinion is that creating these directories (and
potentially failing to cleanup whenever the app fails) isn't
worth the risk.

Signed-off-by: Ralph Castain <rhc@pmix.org>