OFI:  move OFI_RETRY_UNTIL_DONE to common
In running some MPI stress tests on a HPE SS11 network we
had to fall back to using the OFI BTL path.  That in turn
revealed some places in the BTL where we need to use a function
similar to the MTL_OFI_RETRY_UNTIL_DONE macro.

So as a first step move this macro to ofi common layer and
invoke the more general opal_progress function.  That's the
content of this PR

Additional changes needed to the OFI BTL will applied in subsequent
PRs.

The OFI MTL will require more work as the situation hit with
the HPE CXI provider indicates a need to implement some kind of
send backlog queueing mechanism in the MTL rather than simply
spinning on the OFI CQs hoping for progress at the OFI provider
level.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>