libevent: Update autotools versions
Libevent 2.1.12 is the most recent release of Libevent, but is still
ancient, and therefore was built with an ancient version of the
Autotools.  This change takes the official Libevent 2.1.12 release,
runs autogen to update the versions of autotools, and builds a new
tarball.  This was done with Autoconf 2.72, Automake 1.16.5, and
Libevent 2.4.7.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit b85957ee9bcd40fd3b9b2ef5b7a1e2709f0dd2a1)