docs: Be flexible in tarball names
The version of the embedded hwloc and libevent are parsed out from
tarball names.  Make the regex for the version slightly more flexible
so that if there's a name change like a -ompi in the version, this
doesn't break the regex.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit c79f5540b7384bc8161549538dbaf8ff04d11dfb)