coll/han: disqualify hierarchical scatterv for heterogeneous communicators
Hierarchical *v collective algorithms may not work for heterogeneous
communicators with different endianness, interger representation, etc.,
and thus require knowledge of the global communicator's homogeneity to
disqualify the module.

The hierarchical scatterv algorithm requires that every process have the
same architecture as the Root due to the use of MPI_BYTE on node
leaders. Heterogeneous communicators need additional logic to correctly
pack and unpack the data at the cost of memory usage and performance.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>