comm/cid: fix for edge case with intercomm creation
A barrier was removed from the cid allocation algorithm
for intercommunicators that leads to random hangs when
doing MPI_Intercomm_create or moral equivalent.

See issue #12367

the cid allocation algorithm and comm from cid lookup
was modified to use OMPI_COMM_SENTINEL as the placeholder
in the list of pointers to ompi commucators.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>