Utilize PMIx type checking for PML check
It appears that sometimes we get a bogus value back
for the PML selected by a remote proc. Unfortunately,
the current code for exchanging that info removes all
type information from the exchange.

Replace the use of those macros with native PMIx
calls so we can check that the data type being
returned is the one we expected. True, it should not
have changed - but this will help detect and debug
any errors.

Signed-off-by: Ralph Castain <rhc@pmix.org>