Remove the SM collective component
It has been pointed out that the SM component was turned off for a very
long time, so it was not used and not tested. Moreover, when enabled the
performance are below tuned, so there is no reason to keep it around.

The good news is that there are two possible replacement that are up to
review and merge (ACOL and XHC), so at the developers meeting in Austin
2024 we decided to remove the SM collective component.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>