btl/ugni: remove ugni btl and related doc notes
Note all references to uGNI are removed from the docs as
some are referenced via UCX support, etc.

We are removing this as we no longer have access to systems
using Cray legacy gemini and aries networks where uGNI was used.

An entry will be needed to the yet-to-be started changelog for v6.0.x

Signed-off-by: Howard Pritchard <howardp@lanl.gov>