Add bucket allocator for device memory
This allocator is based on the existing bucket allocator but stores
the bucket information separately instead of in front of the allocated
memory chunk. The allocator uses a hash table to store the metadata and
caches allocations. By default, chunks between 4K and 1G are cached.
MCA variables can be used to configure these thresholds.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>