opal/mca/ofi: select NIC closest to accelerator if requested
This patch introduces the capability to select the closest NIC to the
accelerator device. If the accelerator or NIC PCI information is not
available, fallback to select the NIC on the closest package.

To enable this feature, the application should set the MCA parameter
OMPI_MCA_opal_common_ofi_accelerator_rank(default -1) to a non-negative
integer, which represents the process rank(0-based) on the same
accelerator.

The distance between the acclerator device and NIC is the number of
hwloc objects inbetween, which includes the lowest common ancestor on
the hwloc topology tree.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>