common/ompio: add support for gpu aggr. buffers for read_all
add support for using accelerator buffers in the aggregation step of the read_all operation.
This is in common/ompio instead of the fcoll component, since all fcoll components (except individual)
use at the momment the default implementation, which has been moved to common/ompio a while back to
avoid code duplication.

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>