Generate interfaces for C with bigcount
This adds scripts for generating the Fortran and C API bindings from
template files, while also generating bigcount interfaces for those
that require them. The Fortran binding code (for mpi_f08) is designed to
generate both the Fortran subroutines and wrapping C code which
calls into the MPI C api.

Python >=3.6 is required for running these scripts, which is only
necessary when the binding files have not already been generated.
Users of the distribution tarball should not need to generate these files
and thus should not require Python.

Co-authored-by: mphinney1100 <mphinney@lanl.gov>
Co-authored-by: Howard Pritchard <hppritcha@gmail.com>
Signed-off-by: Jake Tronge <jtronge@lanl.gov>