Generate interfaces for C with bigcount
This adds scripts for generating the C API bindings from template files,
while also generating bigcount interfaces for those that require them.
The binding script also include initial support for the mpi_f08 Fortran
bindings, but doesn't yet make any changes to fortran/use-mpi-f08

Python >=3.6 is required for running these scripts, which is only
necessary when the binding files have not already been generated.
Users of the distribution tarball should not need to generate these
files and thus should not require Python.

Note that there are additional changes needed to Open MPI for bigcount support.
One is that the datatype system needs to be embiggened.  This
PR includes workarounds for this lack of support that will need to
be changed once the datatype system is embiggened.  These areas
of code are marked with TODO:BIGCOUNT.

Co-authored-by: mphinney1100 <mphinney@lanl.gov>
Co-authored-by: Howard Pritchard <hppritcha@gmail.com>
Signed-off-by: Jake Tronge <jtronge@lanl.gov>