Generate mpi_f08 bindings and add CFI support
This updates fortran/use-mpi-f08 to generate most of the Fortran
bindings from a script and template files. It also adds support for
Fortran TS 29113 when possible, allowing for better Fortran array
handling that matches the standard.

The C files were imported from PR #10302 and converted to templates to
be fed into the binding script.

Co-authored-by: Gilles Gouaillardet <gilles@rist.or.jp>
Co-authored-by: Howard Pritchard <howardp@lanl.gov>
Signed-off-by: Jake Tronge <jtronge@lanl.gov>