btl/uct: add support for using an another memory domain to form connections
The UCT BTL looks for a connect-to-iface interface in each memory domain to
form connections for connect-to-endpoint transports. For example, with ib the
btl will pick the UD transport as the means to setup RC. While there are
connection transports available (RDMACM) I chose using UD (etc) to support
networks that did not necessarily provide a connection transport.

I am currently working with improving support for Open MPI on a RoCEv2 system
that does not provide support for UD (yet). This breaks the assumption that
there will always be a connect-to-ifact transport available in all memory
domains. To fix this issue this change updates the detection logic to locate
a suitable transport for making connections (tcp by default). If a memory
domain does not have a suitable connection transport the alternate will be
used instead. This has been tested on our broken-UD system and works well.

It a connection-only transport is not needed the extra transport module is
destroyed and the in-memory domain connection transport is used.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>