btl/uct: fix fetching atomics support for osc/rdma
The check that requires 32 and 64 bit atomic support is way outdated at this
point. This commit takes the union of the two instead of the intersection
because technically it is up to the btl user to determine which of each type
are supported. This commit also ensures that MCA_BTL_ATOMIC_SUPPORTS_32BIT is
set when 32-bit atomics are available and ensures that the required
MCA_BTL_FLAGS_RDMA_REMOTE_COMPLETION flag is set on the btl.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>