Protect UCX ep against concurrent creation
If multiple threads initialize the same destination ep simultaneously we
endup in situation where we either segfault or we use resources that we
will never release (because they are overwritten by another thread).
Thus, protect the ep creation with an atomic update, allowing a
single thread to create the ep for a particular peer.

Fixes #12833.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>