sessions: fix some problems with multiple init/fini
Fix a couple of problems uncovered in issue #12854.

Turns out the MCA param management system was "remembering" things
even if a variable was deregistered when a framework was closed.

Also the test case showed that destructing ompi_mpi_session_null
needs to be moved to ompi_mpi_instance_release.

Related to #12854

Signed-off-by: Howard Pritchard <howardp@lanl.gov>