coll tuned add version identifier to the rules file
the version identifier is optional but when provided it  must have the
following format and must appear on the first line.`rule-file-version-N`
where N is an unsigned integer. Older versions of the
parser will fall back to fixed decision mechanism when this line is
present. Version 1 is the original format, Version 2 has support for
optional coll_tuned_alltoall_algorithm_max_requests specification.

Signed-off-by: Burlen Loring <bloring@nvidia.com>