Allow the dpm to query the server for local peer info
The local PMIx server may have info on the local peers
for a job - so allow the client to request that info
if it doesn't already possess it.

Also clean up some diagnostic output so it shows the
PMIx attribute name (or just the provided string if it
isn't an attribute name) to make it easier to correlate
to the code.

Signed-off-by: Ralph Castain <rhc@pmix.org>