ci: Fix build of branch commits
The community Jenkins jenkinsfile is evaluated both on each PR and
for each commit to one of the branches in the openmpi/ompi repository.
The builds for the commit checks were broken because of a bug in the
description setting code, that assumed every invocation was for a PR.
This commit fixes that by differentiating the type of build before
setting the description.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 58cde2fd1d7f378417d813abb63c2134175335ca)