Tracking framework for xpmem rcache registrations in acoll.
A hash table, as part of the acoll modules struct, is used to track the
rcache registrations done as part of the register_and_cache api called
from acoll collective components. This hash table is then iterated over
during module destruct and each rcache registration is deregistered to
ensure that the rcache module destroy proceeds correctly.

Signed-off-by: Mithun Mohan <MithunMohan.KadavilMadanaMohanan@amd.com>