coll/accelerator: allow to select functions to register
This PR introduces the ability to register the component only to the
select functions specified by an MCA parameter string. The idea and the
code is based on the UCC component, and some of the bits might be moved
later to coll/base to make the mechanism more gnerally available to
other components as well.

Note, that the PR introduces the define statments for all MPI collective
operations, not just the ones support by the component at the moment,
since it is a bitmask based operation, and we anticipate to add support
for more collective operations into coll/accelerator shortly

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>