comm: couple of enhancements wrt pmix groups
add a pmix timeout option for group operations.
This may be a no-op with certain pmix server variants but set it anyway.

For protection make sure when creating intercomm communicators using
pmix group construct that all procs supply the same ordered list of pmix procs
to pmix group construct.

update the description for ompi_mca_mpi_pmix_connect_timeout to not
it can be used to control timeout for pmix group calls as well.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>