btl/uct: fix default exclusivity (was too high)
The uct btl should always loose to btl/sm for shared memory peers. This requires
that the exclusivity be set below MCA_BTL_EXCLUSIVITY_HIGH (the exclusivity of
btl/sm). The uct btl had its exclusivity set to MCA_BTL_EXCLUSIVITY_HIGH which
is incorrect. This commit updates the exclusivity to the intended value.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>